Developed a new, alternative, cleaner, and more elegant solution to the following problem:
   ** VERY IMPORTANT NOTE: Any subroutine/function that uses the expression tree as e.g., apply or apply_fun, STILL does
   not work PROPERLY, as currently GuardTemp() only increments the memory guard counter for the ROOT of the tree.
   This issue has to be investigated further.

The original solution was proposed in commit 0302b078ede5c81fca879a1e738627c37a7f17fc, and required three non-deferred
type bound procedures to be added/implemented in class(base_operator). The new solution does NOT require any new method.
Any intermediate temporal node in the expression tree will get a 2 as its reference counter. The root of the tree will
get a 1 if it is temporary. Any method that only uses the expression tree (e.g., apply, or apply_fun), will not deallocate
any intermediate node as far as the reference counter on entry for the root is larger or equal to 2, or if it is permanent.