I have included the discontinuous Galerkin case.
* In create_global_dof_info.f90 I have also considered the coupling by face integration
when creating the dof_graph.

* In fem_space_create.f90, I have made some modifications, in order to be able to use
the same subroutines in the parallel context, by introducing an optional argument,
num_ghosts, that is set to zero for the serial case, and is used in the allocation of
the list of elements (num_elems + num_ghosts in parallel)

* New par_create_distributed_dof_info.f90, that creates the elem2dof and object2dof and
dof_graph. It also includes in the local problem, additional dofs on interface faces due
to discontinuous Galerkin. These dofs are not generated for serial problems, only in the
parallel context, which has motivated this subroutine. It is needed e.g. for dG BDDC.
Next, we call the dof_graph, and the graph_distribution_create (parallel only gluing info).