Intensive work on block_dof_distribution_create.f90.
At this point the subroutine does not work yet, but I think I cannot go further till we solve the following issue:

For dG methods, one DOF can be in more than one object, and all objects are of face type.

As a result, the size of the array to create objects, lst_parts_per_dof_obj, is not the number of dofs in the interface, it is larger than that. It was not like this for cG.

For cG, since every DOF belongs to only one object, the size of lst_parts_per_dof_obj and l2ln2o are identical, and when we sort lst_parts_per_dof_obj we get l2ln2o.
This way, nodes are numbered as 1) interior 2) interface, and on the interface, 1) object 1, ... , n) object n

For dG, the previous statement has no sense, since one node belongs to more than 1 object.

I leave it for discusion...