Further modifications in block_dof_distribution_create.f90
I have replaced l2ln20* with dofs_object*

because it is what we have now.

NOTE: For cG it is the same so everything works as before

For dG, dofs_object is an array of size sum_objects(dofs_object)
*(remember that a DOF can be in multiple proc's and the size of this array exceed ndofs)
**(note the memrealloc when it happens, not for cG).

This array is sorted with the table lst_parts_per_dof_obj.

Now I am filling properly this array for dG too.

Next step:

1) Cut into "objects" *lst_parts_per_dof_obj* and *dofs_object*

This process gives us the objects as aggregation of DOFs, as we want. It will be right for cG, dG, and cdG.

2) Create import by aggregation of objects (verbatim from what we have?)

3) Replace the lobjs pointers with the dofs_object split into parts everywhere (remember that DOFs in an object are not consecutive)

IMPORTANT QUESTION: Can we ELIMINATE the reordering? Now we cannot have anymore DOF ordering: Interior, Object 1, ..., Object N

But we can still have: Interior, Interface.

Can we omit even the interior 1st interface 2nd order?

Not a critical point...