Added the _t postfix to all derived data type names.
This process was quite time consuming. It could only
be partially automatized (manual adjustments were required
afterwards, e.g., when the name of an instance or dummy argument
is a substring of the data type name, or there are two data
type names s.t. one is a substring of the other. Besides type
names can appear in multiple contexts, e.g., when allocating an
allocatable of polymorphic data types, in an import or public
declaration, with the class or type attribute, etc., which significantly
complicates automatic parsing/replacing. If u have an idea of how
this process can be fully automatized, please tell me.

See the following page:
  http://servercomfus/LSSC/fempar/wikis/coding_style_guidelines
  in GitLab's WiKi.