Workaround due to a BUG in the IBM XLF v14.1 compiler.
- BUG REPORT:

Dear JSC support staff,

we believe that we found a potential BUG in the IBM XLF2008 compiler
installed on JUQUEEN. In particular, we observed that the compiler does
not perform default initialization of derived data type components when
polymorphic allocatables are allocated (i.e., via the mold= keyword).

The same issue was reported (and solved) for the GNU compiler (gfortran)
some time ago. Please see the following post:

https://gcc.gnu.org/ml/fortran/2010-06/msg00133.html

In particular, the following paragraph:

<<There is one issue left, which is demonstrated by the last part of the
third test case: Default initializtion does not work yet if the MOLD
expr itself is polymorphic. In this case we obviously cannot determine
the type to be used for def.-init. at compile time (which is why the
current implementation does not work).>>

In order to reproduce the issue, you can compile and execute the test code
available at:

https://gcc.gnu.org/ml/fortran/2010-06/msg00133/allocate_alloc_opt_10.f90

Finally, my recommendation would be to redirect this issue to the IBM XLF2008
developers, taking into account that it seems to be an issue to be solved by
compiler developers.

Thanks in advance.
Best regards,
 Alberto.

- SOLUTION:

    A new procedure (default_initialization) was created to initialize
    the components of all derived data types that inherit from
    memory_guard_t data type.

    Now, after allocate this derived data types via mold its mandatory
    to initialize its components using the procedure default_initialization