Introduced create_id and get_id in migratory_element and made it private.
Introduced ip_element_id_t as an extension of element_id_t containing just ielem.

The element_id_t is now an argument to the abstract triangultion_create. The routine
of the derived class, e.g. serial_triangulation_create calls that of the mother
class (not TBP) it sends the argument to select the type of element id.