Fix some checks in the previous commit to remove out-of-bounds index while sorting and compressing.
New signatures for append_body/update_body generic procedures.
Now it is allowed to append new entries or update existing
entries without specifying the allowed bound where to insert.

When calling append_body procedure, if a entry is out-of-bounds, it
will be removed while sorting and compressing.

When calling update_body procedure, if a entry is out-ouf-bounds,
it will be ignored.