I have created a new branch to modify the treatment of bc's.
Now, the strong_dirichlet_data is a global array (of the subdomain) in the
serial_fe_space_t object.

I have created it (set_up_boundary_conditions in sbm_serial_fe_space.i90) and
free it. I am touching the elem2dof with a negative integer that ( times -1)
allows me to take Dirichlet data from the strong_dirichlet_data array.

@ocolomes: Please, can you fill this array and ELIMINATE bc_code and bc_value.

I have modified the fe_function_t. Now, it includes the nodal_values and
strong_dirichlet_data array. They are allocated and filled in the create (in the
fe_space). I AM COPYING THE VECTORS BUT I THINK IT IS THE WAY TO GO.

It involves modifications in the code. I have modified one of the drivers
by @eneiva accordingly.

@eneiva: Can you merge these changes with yours and see everythins is OK.

After the work by @ocolomes, we will have ready the new bc treatment. It
would solve the problem pointed out by @molm.