Changes acording with comments in merge request !49:
(1) Review the implementation of this subroutine:

    Rename by direct_solver_free_in_stages.
    action dummy argument should not be optional.
    The polymorphic member variable of type class(base_direct_solver_t) should only be deallocated when action==free_clean.

(2) Rename free by free_in_stages.
    Add a new TBP of type(direct_solver_t) called free which calls free_in_stages with action=free_numeric|free_symbolic|free_clean.

(3) I think it is better to try to implement the state transition diagram on base_direct_solver_t, so that we can code re-use.
    Current deferred methods of base_direct_solver_t should become body deferred methods, i.e., primitive methods for the wrapping template method.

(4) free_numerical could also be called when this%state_is_symbolic() or this%state_is_start()

(5) free_numerical could also be called when this%state_is_start()

(6) Eliminate set_defaults from type(direct_solver_t).
    To think whether it makes sense to have a deferred method of class(base_direct_solver_t) that returns a type(ParameterList)
    initialized with all parameter-value combinations to default values. (may be for a later stage).

(7) We agreed that direct_solver_names should only publish the public symbols it declares locally to avoid additional couplings.
    Each implementation should publish its own public symbols. serial.f90 will use direct_solver_names, base_direct_solver_names,
    and any_implementor_names

(8) Be careful with the refresh of these member variables when going back to BASE_DIRECT_SOLVER_STATE_START,
    e.g., after an update_matrix call with different nonzero pattern

(9) To check whether the user set a matrix type via parameter list or not. In the latter case, the type of matrix should be extracted from the matrix itself

(10) Be sure that we issue a warning if we try to change the matrix_type with symbolic or numerical already computed.

(11) Add update_direct_solver_matrix(direct_solver, same_nonzero_pattern) TBP of type(affine_operator_t)