Undo the following work from commit 61efa6dfd304961ebc2a02ac0e2182153e0f04f2:
   <<Modified function.f90 so that all superclasses declared there have now a number_dimensions member variable
   which determines the number of space dimensions in which the corresponding function leaves (This is something
   that we neglected in issue #91). I am quite sure that letting the implementors know the number of dimensions will
   allow to write significantly safer code, and avoid code replication. Note that a generic "create" TBP is declared in these
   superclasses, and overloaded with a TBP that just initializes number_dimensions member variable. If subclasses need additional data to be created
   (see e.g., type(constant_vector_function_t)), then the may overload "create" with a specialized TBP on the subclass.>>

Therefore, if we are writing the code that evals a *2D* type(vector_function_t) user-defined
function, and the code has been compiled with "integer(ip), parameter :: number_space_dimensions = 3",
then we must ensure (to avoid problems when evaluating FE computations) that the third component
of the type(vector_function_t) instance returned by the function is 0.0_rp.

On the other hand, if we are writing the code that evals a *3D* vector function, and we need to access either to the z
component of the point, or to write the third component of the type(vector_field_t) instance (something which will be required for any non-trivial 3D function), then the code
will crash unexpectedly with a runtime error (hopefully) handled by the compiler RTE (i.e., cannot be handled by FEMPAR)
if FEMPAR has been compiled with "integer(ip), parameter :: number_space_dimensions = 2".

Finally, NOTE THAT a 3D function might not be reused "as is" for solving a 2D problem,  as it might not satisfy the requirement
above on the third component of the type(vector_function_t) instance being 0.0.