Skeleton of the implementation (object, methods, testing driver)
Created the new type tri_lagrangian_reference_t. At the moment, I just copied
the object and methods of quad_lagrangian_reference_fe_t and replaced 'quad' by 'tri'.

I had to rename most of the low-level subroutines, otherwise they would appear
duplicated, preventing compilation. This is a temporary fix to force compilation.

@sbadia @amartin @ahierro ->  Currently in FEMPAR, quads/hexas are referred as 'quad',
whereas tris/tetras as 'tet'. This naming convention is not consistent from the
dimensional point of view. Is there any reason for this? Or should we rename it?