Removed topological-dependence of set_nodal_quadrature()
Transformed fill_nodal_quadrature() into a TBP. At the moment, this is a little
dirty because I pass to the quad version the unused argument 'order'. However,
I think that it is possible to have the same fill_nodal_quadrature() using TBPs
to get the reference coordinates of the element (as with tris).