volume/face_integrator get all values from interpolation_o_map
Up to now, for discrete integration and evaluation of cell/face FE functions,
shape function values were extracted from the interpolation in the reference space,
whereas shape function gradients were extracted from the interpolation in the
physical space.

This relies on a property that is particular of lagrangian finite
elements (shape functions values are invariant by the FE map) and
precludes extensibility to other type of finite elements.

Moreover, there is the discussion in 4854bc85. The interpolation_o_map in faces
is created from the interpolation of the first face. When updating the integration,
the shape derivatives of the active face in the physical space are updated, but the
shape values are still the ones from the first face.

From now on, all volume/face_integration get_values/gradients and evaluate_fe_function
procedures extract the values from the interpolation_o_map (the natural and extensible
way). This commit contains all the necessary changes for this. The most important one
is that now `interpolation_o_map` member of `interpolation_face_restriction_t` is no
longer a single instance of `interpolation_t`, but an allocatable array, that is the
counterpart of member `interpolation(:)` in the physical space.