Finished all the machinery to generate geometrical objects for an arbitrary topology (geometry_tree_t),
Children iterator over the previous array (children_iterator_t)

Idem for the generation of nodes and their relationship with geometrical objects (node_array_t)

Plain iterator over the previous array (node_iterator_t)

Now, the fill_scalar function in reference_fe_t is now implemented in lagrangian_reference_fe_t, instead of
hex and tet implementations

*IMPORTANT*

For tets, there were some errors in the previous version of the code. 'orientation' and 'vertices_vef' were
not correct. 'vefs_vef' was not implemented.

There is a swap beetween 7 and 8 objects in TET (3D) but it should not affect the code.

I have compared the two versions and they provide the same up to the previous comments BUT not tested for
a simulation.

*EXTREMELY IMPORTANT*

Now, 'vefs_vef' meaning has changed, it only includes children of one dimension less. It requires a
modification of the code developed by @principe .