Several BUG-fixes applied to RT FEs implementation resulting from commit ID 633f522a7baf45638cd3e69d120b41b6942c389c. In particular, to raviart_thomas_apply_femap_to_interpolation. The other two smells are still present, and still to be solved.
In the current status, the coefficient linear system is simmetric indefinite
(and therefore non-singular). However, when imposing p=g=1 on the whole
boundary of the domain, although p=1 in the whole domain, u = -qrad(p) is only
zero for some of the velocity DoFs. I agreed with @sbadia that a simpler
validation test with order=1 RT FEs should be developed.