Implemented read and write in par_environment. Now the partitioner writes num_levels and the arrays num_parts_per_level and parts_mapping for all tasks involved in a multilevel simulation. Actually only parts_mapping is variable, the rest are the same for all tasks.
The creation of the environment using these functions (insted of getting
them from the driver is NOT developed (yet).

Implemented the iteration over the FPL to add parameters to CLI and to
get them after parsing. This is a prototype to FPL and FLAP coupling.
It is done in the driver part.