Refactored (moved) communications functions from par_environment to par_context. Now par_environment is independent of MPI and relies on par_context only.
Implemented the creation of par_context (again) independent of triangulation but
still transparent to the user, either reading files or calling the uniform mesh
generator. There is still some (minor) refactoring to be done here to eliminate this
dependency.

Eliminated the intercommunicator l1_to_lgt1 and implemented the bcast in world_communicator
by performing a send/receive from 0 to n and then a bcast in lgt1.