Modified par_context_t to make it abstract.
Introduced mpi_context_t and serial_context_t.

Renaming of par_environment into environment still pending; this is
the next operation and this commit is a backup, I did not run tests.

Modified creation of serial triangultion to include the creation
of the environment (still called par_environment) as a serial one.

Modified creation of environment to handle either serial or mpi
contexts according to parameter execution_context (given in the FPL).