Elminated serial environment, unused since last commit.
Eliminated abstract class environment.

Renamed par_environment into environment. Now there is only one
class environment_t.

The creation of the environment is performed inside triangulation.
In the parallel case there is still the option of creating the
environment in the driver and send it to the triangulation, it
is an optional argument. In the driver par_test_poisson this
option is not exploited.