Reverted the following changes from commit id 010925f83d79d43ecaa75ba0cac9ca7d4abf7859
    - Bug-fixed sbm_par_triangulation.i90 (uniform mesh case). My current fix to this bug involves
      a cell-based nearest neigbour exchange of cell_gids right after the cell_import is set-up.
      This exchange fixes the l2g ordering of ghost cells. I think that this
      exchange is required provided that a given subdomain cannot determine (at least without comms)
      where a given ghost cell is located in its neighbouring subdomain.

After speaking with @principe, the BUG-FIX is significantly more complex and requires careful thinking.
The issue is caused by the fact that cell_import%create() generates a local2global correspondence for ghost
cells which is not consistent with the one generated by uniform mesh generator. In the unstructured mesh
itinerary, there is a single local2global correspondence (the one built during cell_import%create()), this is
why it works for unstructured meshes.