Introduced mpi_omp_context, an hybrid implementation of execution_context_t
Introduced several modification to make the code run using OMP:
- eliminated omp in hash tables as they are now threadprivate
  (now we are exploiting coarse grained parallelism)
- introduced omp_thread_private clause in many places
- introduced critical in io_open which inquieres before opening,
  otherwise a race condition appears.

WARNING: in this commit the CLA is NOT used. I commented the lines in
parameter generator because of the error associated to deferred-length
characters. This commit is backup of the work done and an intermediate
step to switch to the intel compiler.