Solved the following execution issue raised by par_test_poisson:   http://servercomfus:8080/testDetails.php?test=1952&build=10275
It was caused by the fact that two TBPs of class(base_mlbddc_t) were
not annotated with the "recursive" statement. These two TBPs call
to themselves INDIRECTLY (i.e., A calls B, B calls A). The program
to be standard conforming still requireis the recursive statement to
be specified in the declaration of A. The most strange think is that
the issue was only revealed without -qopenmp and linking against the
bunch of dynamic libraries of MKL (instead of against its single dynamic
library -lmkl_rt).