Reversed the loops in cells and fields in `serial_fe_space_set_up_strong_dirichlet_bcs`
Now, the outer loop is in fields and the inner in cells.
It used to be the outer in cells and the inner in fields.

With this change we consistently mark the vefs as visited in a per field
basis, avoiding to add an extra dimension to the visited vefs array.