We set as corners all the "problematic" dofs belonging to coarse edges.
A "problematic" dof is a dof that belongs to at least 1 cut
element (taking into account both local and ghost cells).

This is slightly more conservative than the strategy described
in the paper (variant 2, the one that works well for Dirichlet BCs), but
it is almost equivalent in practice (and much more simple to implement).

TODO: The code does not run yet. A FORTRAN runtime error fires up when
creating the coarse fe space.