Extended the `uniform_hex_mesh_t` to create meshes with custom domain limits
Until now the mesh generator used to build a mesh of the domain [0,1]^d

Now it is possible to mesh the domain [x0,x1] x [y0,y1] ( x [z0,z1] )
where x0, x1, y0, y1, ( z0, z1 ) are arbitrary user-defined limits of
the domain in each of the Cartesian directions.

To do so, the mesh generator accepts now a new parameter in its parameter list
named `hex_mesh_domain_limits_key` whose value is a rank 1 `real(rp)` array of length
at least 4 in 2D and at least 6 in 3D.
The domain limits are sorted in the following order within the array:
[x0,x1,y0,y1,z0,z1]. Note that only the 4 first positions will be read
in 2D and only the 6 first positions in 3D. Any other values are ignored.

If the new parameter is not provided. The old behavior of the mesh
generator is obtained.