Weighting the matrices in the mkl pardiso direct solver
WARNING: This is a very experimental commit to check if a Jacobi weighting
helps to solve with pardiso the very ill-conditioned linear systems associated with
unfitted methods. This commit it is very likely to be reverted in the
future. Or at least it will require some code code clean up to be
incorporated in FEMPAR.