Introduced p_l1_coarse_fe_handler_t (a type wrapping a pointer to a coarse_fe_handler) and replaced the pointer to a coarse_fe_handler_t by an allocatable array of pointers (in the same ways as it is done with reference_fe. Modified drivers accordingly.
Modified the interface of coarse_fe_handler TBPs to receive the field_id.
The allocation of the constraint matrix and weighting operator is now made
outside them and they fill the values for each field.

Corrected two bugs appearing in the output handler when using the ifc. The
first of them was a seg fault when accessing cell_vectors(1:0) in a function
call (it is not really accessed) and it was solved allocating it if it is not.
The second was an assumed short circuit evaluation in
output_handler_cell_fe_cuntion_create