Fixed the test_poisson_unfitted_serial for ENABLE_MKL=OFF builds
In this build type, the system is solved with a UN-preconditioned Conjugate
Gradient solver, which is very sensitive to the condition number.

The test fails for order == 5, since it leads to an
extremely ill-conditioned system, the un-preconditioned CG solver has serious
problems, and the tolerances are not met.

Using order 3 instead of 5, circumvents the issue.

TODO: Maybe we can use a preconditioned conjugate gradient solver (instead
of an un-preconditioned one), e.g., with a Jacobi preconditioner, as a
fall-back solver when ENABLE_MKL=OFF. And then, using order 5 again.