New treatment of facets for void/full elements.
Now, it is the user the one dealing with the active cell id, instead of the library. It has required extensive changes with respect to @eneiva solution.

In particular: we have now methods to ask whether a face is in the interior or the boundary of a field domain, as well as whether it is on the subdomain of the full domain.