Intensive changes to handle nonlinear operators
I have created a new fe_nonlinear_operator_t object, which generalizes the previous fe_affine_operator_t. In particular, it provides the possibility to set the point in which we want to evaluate the operator. Further, one can compute the tangent or the residual.

It has also motivated a quite deep change. Now, it is the fe_space_t the one that handles the dof numbering at the create TBP, unless stated otherwise. In particular, for parallel and unfitted fe spaces, we want to call the serial_fe_space_t create TBP without dof numbering computation, in order to make it at the subclass.

We have thus eliminated this from the fe_affine_operator_t (also nonlinear one). Furthermore, the field_blocks and field_coupling must be passed to the create TBP of the fe_space if sthg different from monolithic is required (NOT TO THE OPERATOR).

I have also included new methods in discrete_integration_t to compute the residual and tangent.

I have slightly changed the nonlinear_solver_t by Javier.

All tests but one unfitted test pass.