I have created a new type, linear_solver_t, which only has a method, is_linear, but includes the iterative_linear_solver_t and direct_solver_t.
On the other hand, iterative_linear_solver_t is now a concretization of operator_t, which was not before. To do this, the solve TBP is now the apply TBP of operator_t.

I have included in operator_t a new method, update_matrix, that gives a signal (only sense for linear_solver_t), informing that the matrix has changed. The old update_matrix TBP is now replace_matrix in direct_solver_t.