Bugfix in the way the constraints are applied to the element matrix and vector
The TBP apply_constraints and impose_dirichlet_bc_non_conforming didn't
work for num_fields>1. The problem was that the TBP assumed that the
local dof numeriation of each field expands from 0 to num_dofs in this field.
This is only true for the first field. In the second and remaining
fields the local dof numeration does not start
from zero but from 1+offset, where offset is the total number of dofs of
the preceding fields.