Added a new method, scale_add_fe_operator_t, that, given two fe_nonlinear_operator_t A and B and an integer alpha, it implements alpha*A+B.
We note that it cannot be done with operator algebra within FEMPAR, since we need to explicitly have the entries of the alpha*A+B matrices.

I have to say that I am not using this method for time integration, but it is a simplified version of the one being used in time integration.