Bug-fixed pfs_compute_max_part_id_my_part_id_and_dofs_ggids_field.
In particular, the algorithm was assuming that, on a given processor,
it was enough to fill the my_part_id of a interface vef on its ghost cells
around ONLY for interface vefs which are solely at ghost cells. It turns out that this is
actually not enough, and this info must be filled for all interface vefs, even if
these vefs are at least at one local cell.