Eliminated the following long-standing error-proneness missing feature of type(mlbddc_t):
  "Properties of type(mlbddc_t)%coarse_grid_matrix have to be set-up by L2 tasks. These tasks do not know
   which are the properties of the sparse matrices on L1 tasks. From where should L2 tasks extract these properties?
   Should they be provided by the user via FPL ? Should they be gathered from L1 tasks? Temporarily workarounded by
   providing a fixed set of properties, but clearly not the way to go ..."

A new sub parameter list with key=mlbddc_coarse_matrix_params can now be used by the user to explicitly
define the structure (sign, storage, symmetry) of the coarse-grid matrix to be solved at
the next level in the hierarchy. This sublist is optional. In the case it is not present, then
L1 tasks transfer to L2 tasks which is the structure of the coarse-grid matrix.

See par_test_poisson test for an usage scenario