Bug-fixed improved generation of type(dof_import_t) s.t. it now passes ALL tests.
The code was (erroneously) assuming that an interface coupling that can only be inferred from a
neighbour of a nearest neighbour MUST involve necessarily a processor that is not
nearest neighbour. It might be possible that the former (i.e., the neighbour of a nearest neighbour) is
indeed a nearest neighbour provided the "odd" mesh distributions resulting from space-filling curves.