Fixed failed tests with "Intel_RELEASE_MKL=ON" build reported by CDash at:    http://servercomfus:8080/viewTest.php?onlyfailed&buildid=17454    http://servercomfus:8080/viewTest.php?onlyfailed&buildid=17449    http://servercomfus:8080/viewTest.php?onlyfailed&buildid=17448
The issue was in particular introduced in commit ID 0cc4c900. There seems to be
an INTEL compiler BUG related to array pointer dummy arguments declared as intent(out).
In particular, the array targeted by the actual argument to these sort of dummy arguments
is truncated/polluted with rubish right after the subroutine/function is called. The work-around
is to declare the dummy argument as intent(inout).