Update Gitlab CI scripts in order to tackle a very annoying issue that was revealed after 684932ba018f32b057b8edb34e4276a73925059b
After 684932ba018f32b057b8edb34e4276a73925059b, ctest command was in some cases (seems to be
completely arbitrary) generating a nonzero output error code even when ALL TESTS WERE PASSING.
Gitlab CI scripts are executed in a shell environment where the first command that returns a
nonzero output code causes an abort of the whole script. This was causing two undesired behaviours.
(1) An e-mail was sent to the author of the commit even if the tests were passing (false positives).
(2) If the first ctest call (i.e, that in charge of serial tests) was returnig a nonzero output code, the script
was aborted and the parallel tests were not performed.

I have solved the issue as follows. Using set command before and after the call to ctest, I avoid
that the scripts aborts when ctest returns a nonzero output code. On the other hand, I do not trust
the error code returned by ctest, but check for the error logs generated by ctest (FailedTests.log)
in order to decide whether the test failes or not.