Solved BUG exposed by par_pb_bddc_poisson reported by @hieucsd in issue #238
Since merge request !202, not all faces of the local portion of the mesh are introduced into
serial_fe_space_t%facet_gids, for reasons made clear by inspecting the code there

This was causing trouble to the output produced by par_fe_space_compute_faces_object, in particular,
as this invariant is no longer fullfilled, the expression

-      if ( fe_vef%is_facet() ) then
-        facet_gid = fe_vef%get_gid() - this%triangulation%get_num_vefs() &
-                                    + this%facet_gids%size()

was returning negative facet_gid numbers, which obviously are NOT allowed.

On the other hand, par_fe_space_compute_faces_object was assuming that the GiDs of vefs within a subdomain
were organized in increased order by dimension, and this is not longer true for p4est_triangulation_t. In the
path to eliminate this invariance-depedence, I also develop it in order to take care of hanging and not hanging
VEFs in the presence of non-conforming meshes