Solved a BUG in the interplay among "p4est_bt_extend_p4est_topology_arrays_to_ghost_cells" and "p4est_bt_update_lst_vefs_gids_and_cells_around".
The algorithm in p4est_bt_update_lst_vefs_gids_and_cells_around needs to know whether a hanging edge
is either hanging on a coarser edge or coarser face. This can be differenciated by looking at the hanging
edge edge neighbour. If it exists, then it is hanging on a coarser edge, if it is -1 (i.e. there is no
neighbour) then it is hanging on a face.

In a parallel distributed-memory context, the situation is more tricky. An edge may be hanging on a coarser
edge, but the hanging edge edge neighbour may not exist if it is not in the local portion of the processor
(i.e., local cells + a single layer of ghost cells). Thus, we cannot use -1 (the value used so far before
this commit), but a different flag value to signal this scenario.

The work undertaken in this commit modifies the interplay among  "p4est_bt_extend_p4est_topology_arrays_to_ghost_cells" and
"p4est_bt_update_lst_vefs_gids_and_cells_around" such the aforementioned goal is fulfilled.