Solved CMake configuration error pointed out by Gitlab CI at:   http://ci.servercomfus/projects/2/builds/19254
In particular, in commit 8477fa50e7476021dacf6ecc3490b0c7875d8386, cmake
was generating an error if the CXX was not being explicitly passed in
the invokation to FEMPAR's SuperBuild CMakeLists.txt. In order to
overcome this issue, now both SuperBuild and ThirParty CMakeLists.txt
have been declared as CXX projects as well, so that a default CXX
compiler is searched for.