Added the -fimplicit-none and -implicitnone GNU and INTEL compiler flags, resp., to the Gitlab CI test system
With these flags, if there is a missing implicit none within
the local scope of a program unit (program, subroutine, module, etc.),
then the compiler will automatically raise an error if there are variables
which are declared implicitly. This is in general very dangerous if one
does not know what is doing.