Vast refactoring of unfitted_boundary_cutter_t (abstract) and marching_cubes_t (type extension of the former)
The objective of this refactoring has been to: (1) simplify the signatures of the procedures of the
abstract interface of  unfitted_boundary_cutter_t, and (2) remove from the abstract class
low-level implementation details of marching_cubes_t that we do not want to expose at this level.

(1)

In my view, unfitted_boundary_cutter_t is conceptually seen as follows (perhaps we may
think of a different/better conceptual definition). It is a mesh-like container
able to describe the subtriangulation of cut cells. However, as its clients are cell iterators,
it has to become aware of the current cell on which the iterator is positioned to resolve the queries
related to the current cell.  I have made this concept explicit in the abstract class, i.e., the client
has to signal explicitly via a call to "set_current_cell" where it is currently positioned.
Thus, for any of the deferred procedures it is NO longer needed to explicitly pass as a dummy argument the CELL gid
on which I want to resolve info about the current cell subtriangulation. In the previous version
of marching_cubes_t, there was a hard to define gumbo soup. When calling update_sub_triangulation, it positioned
into the current cell, but on the other hand, it could resolve queries of any other cell, but only for some TBPs,
and not for others ... there was no consistency nor uniformity in the way this was treated.

(2)

I have removed the "update_sub_triangulation" concept from the top-level, abstract classes
(i.e., unfitted_boundary_cutter_t, cell_iterator_t, vef_iterator_t, etc.) This is a very low-level detail,
related to an optimization which is done within marching_cubes_t in order to avoid storing the subtriangulation
of all mesh cut cells at once. In my view, we should not expose this low-level detail and propagate it thorough
the whole project as it was being done so far. Now, update_sub_triangulation is something private, local to the scope of
marching_cubes_t, thus hidden as any other low-level implementation detail.

Besides, marching_cubes_t was developed such that their clients must explicitly call update_sub_triangulation in order
to update the sub_triangulation data structures for the current cell. This caused calls to update_sub_triangulation to
be propagated across the whole project, and fragility, as one way try to access to an un-updated subtriangulation of a previous
cut cell (this situation was being actually detected by means of asserts and thus only in DEBUG mode). It was not clear at all which queries required the subtriangulation
to be updated.

After the refactoring, I have followed a "deferred" update_sub_triangulation strategy. I.e., whenever we move to a different cell,
we flag that sub_triangulation is no longer updated, i.e., we dont actually update it at this time. The update is deferred until
a given procedure of marching_cubes_t requires it to be updated.