After discussion with @fverdugo, revisited some of the design decisions regarding unfitted_boundary_cutter_t (made in commit 11e31d98f66648b3af4c72e6cf9ca1a4b67d224b)
In particular, removed the "set_current_cell" concept from unfitted_boundary_cutter_t.
The latter is now able to provide data related to the subtriangulation of ANY cut cell
(i.e., no longer only the current cell on which the client is positioned). In order to
do so, all signatures of the deferred procedures of unfitted_boundary_cutter_t now take
as input argument the cut cell on which we want to query anything related to its subtriangulation.