Fixed an issue in refinement_strategy_names module that has appeared after merge request !289
In particular, the data type within that module was designed such that the user is assumed to get the
refinement_and_coarsening_flags from the triangulation and pass to it afterwards.

This is no longer the way to go. Refinement and coarsening flags MUST BE PROVIDED via cell_iterator_t