In this commit I have developed the necessary code in FEMPAR such that the user may play around on the parameter "k" for the 2:1 k-balance and the k-ghost cells (either from the command line handled by fempar_parameter_handler_t and/or setting appropiate (key,val) pairs into the parameterlist_t passed to triangulation%create(...)).
The default values for these two parameters are the most conservative possible (the code will work in all cases).
If you want to modify them, you need to know what you are doing, as not all possible
values are valid for all possible FE spaces. See paper "A generic finite element framework on parallel tree-based
adaptive meshes for more details"

So far there is only experimental support for these two parameters, I only tested k=0,1 for 2:1 k-balance and k=0
for ghost cells. I suspect that k=2 for 2:1 k-balance (e.g., for RT FEs) requires crosscutting devs in the code.
On the other hand, k>0 for ghost cells may work "as is" now in FEMPAR, but I did not try it.